Move GetCSharpDynamicLocalInfo to EE
